RDD.jl Documentation

RDD.jl Documentation

RDD.jl Documentation

Functions

Index

Functions

RDD.AbstractRDDModule.partitions — Method.

Returns the list of partitions of the given rdd.

source

RDD.AbstractRDDModule.preferredlocations — Method.

preferredlocations(rdd, partition)

Returns the list of preffered locations ordered by preference for a partition of an rdd.

source

Index

RDD.AbstractRDDModule.partitions

RDD.AbstractRDDModule.preferredlocations