AbstractRDD

AbstractRDD

RDD.AbstractRDDModule.AbstractDependency

RDD.AbstractRDDModule.AbstractLocation

RDD.AbstractRDDModule.AbstractPartitioner

RDD.AbstractRDDModule.AbstractRDD

RDD.AbstractRDDModule.dependencies

RDD.AbstractRDDModule.iterator

RDD.AbstractRDDModule.iterator

RDD.AbstractRDDModule.iterator

RDD.AbstractRDDModule.partitioner

RDD.AbstractRDDModule.partitions

RDD.AbstractRDDModule.partitions

RDD.AbstractRDDModule.partitions

RDD.AbstractRDDModule.preferredlocations

RDD.AbstractRDDModule.AbstractDependency — Type.

Parent type of rdd dependencies.

source

RDD.AbstractRDDModule.AbstractPartitioner — Type.

Parent type of rdd partitioner.

source

RDD.AbstractRDDModule.AbstractRDD — Type.

Parent type of rdds.

source

RDD.AbstractRDDModule.dependencies — Method.

dependencies(rdd::AbstractRDD)::AbstractVector{AbstractDependency}

Returns the list of parent rdds with a specification of the way this rdd depends on each one.

source

RDD.AbstractRDDModule.iterator — Method.

iterator(
    rdd::AbstractRDD{T}, 
    partition::Int,
    parentiters::AbstractVector
)::AbstractPartitionIterator{T} where {T}

Returns an iterator on a partition of the rdd given parent partition(s) iterators. This method must be implemented by subclasses of AbstractRDD.

source

RDD.AbstractRDDModule.partitioner — Method.

partitioner(rdd::AbstractRDD)::Union{AbstractPartitioner, Nothing}

Returns the partitioner of this rdd. Default implementation returns nothing.

source

RDD.AbstractRDDModule.partitions — Method.

partitions(rdd::AbstractRDD{T})::Int64 where {T}

Returns the number of partitions of the given rdd. This method must be implemented by subclasses of AbstractRDD. The default implementation raises an error.

source

RDD.AbstractRDDModule.preferredlocations — Method.

preferredlocations(
    rdd::AbstractRDD{T}, partition::Int
)::AbstractVector{AbstractLocation} where {T}

Returns the list of preffered locations ordered by preference for a partition of an rdd. The default implementation returns no preffered locations.

source

RDD.AbstractRDDModule.AbstractLocation — Type.

Parent type of rdd partitions preferred locations.

source