MapPartitionsRDD

MapPartitionsRDD

RDD.MapPartitionsRDDModule.MapPartitionsRDD

RDD.MapPartitionsRDDModule.MapPartitionsRDD — Type.

struct MapPartitionsRDD{T} <: AbstractRDD{T}

Type of an rdd whose partitions are derived from a parent rdd partitions by application of a map function.

source

RDD.AbstractRDDModule.iterator — Method.

iterator(
    rdd::MapPartitionsRDD, 
    part::Int, 
    parentiters::AbstractVector{AbstractPartitionIterator}
)::AbstractPartitionIterator{T} where {T}

Implementation of iterator for MapPartitionsRDD. Returns the elements of parent partition transformed by map function.

source

RDD.AbstractRDDModule.partitions — Method.

partitions(rdd::MapPartitionsRDD)::Int64

Implementation of partitions for MapPartitionsRDD. Returns the same partitions count as parent rdd.

source