ParallelCollectionRDD

ParallelCollectionRDD

RDD.ParallelCollectionRDDModule.ParallelCollectionRDD

RDD.ParallelCollectionRDDModule.ParallelCollectionRDD

RDD.ParallelCollectionRDDModule.ParallelCollectionRDD

RDD.ParallelCollectionRDDModule.ParallelCollectionRDD — Type.

struct ParallelCollectionRDD{T} <: AbstractRDD{T}

A subclass of AbstractRDD representing distributed in-memory collection.

source

RDD.ParallelCollectionRDDModule.ParallelCollectionRDD — Method.

ParallelCollectionRDD{T}(col::AbstractVector{T}) where T

Create a ParallelCollectionRDD from a AbstractVector. The number of partitions will be the number of workers of a julia cluster as returned by nworkers(). If there is less than 2 workers, 2 partitions will be created.

source

RDD.ParallelCollectionRDDModule.ParallelCollectionRDD — Method.

ParallelCollectionRDD{T}(col::AbstractVector{T}, numpart::Int) where {T}

Create a ParallelCollectionRDD from a AbstractVector with numpart partitions.

source

RDD.AbstractRDDModule.iterator — Function.

iterator(rdd::ParallelCollectionRDD, numpart::Int, parentiters::AbstractVector = [])

Implementation of iterator for ParallelCollectionRDD.

source

RDD.AbstractRDDModule.partitions — Method.

partitions(rdd::ParallelCollectionRDD)

Implementation of partitions for ParallelCollectionRDD.

source