Transformations

RDD Transformations

RDD.Transformations.flatmap

RDD.Transformations.flatmap — Function.

flatmap(f, rdd::AbstractRDD, type::Type = Any)::AbstractRDD{type}

Given a rdd and a function f that returns an iterator, returns a rdd whose elements are those coming from all the iterators f(x) for x in parent rdd.

source

Base.filter — Method.

filter(f, rdd::AbstractRDD{T})::AbstractRDD{T} where {T}

Returns a filtered rdd from rdd with the given filter function f.

source

Base.map — Function.

map(f, rdd::AbstractRDD, type::Type = Any)::AbstractRDD{type}

Generates a new rdd by transforming elements of rdd with function f. type should be the returned type of map function f.

source