AbstractRDD

AbstractRDD

RDD.AbstractRDDModule.AbstractPartitioner

RDD.AbstractRDDModule.AbstractRDD

RDD.AbstractRDDModule.Dependency

RDD.AbstractRDDModule.compute

RDD.AbstractRDDModule.dependencies

RDD.AbstractRDDModule.dependencies

RDD.AbstractRDDModule.iterator

RDD.AbstractRDDModule.iterator

RDD.AbstractRDDModule.iterator

RDD.AbstractRDDModule.partitioner

RDD.AbstractRDDModule.partitions

RDD.AbstractRDDModule.partitions

RDD.AbstractRDDModule.partitions

RDD.AbstractRDDModule.preferredlocations

RDD.AbstractRDDModule.AbstractPartitioner — Type.

Parent type of rdd partitioner.

source

RDD.AbstractRDDModule.AbstractRDD — Type.

Parent type of rdds.

source

RDD.AbstractRDDModule.Dependency — Type.

Type of rdd dependency.

source

RDD.AbstractRDDModule.compute — Method.

compute(rdd::AbstractRDD, partition::Int)

Compute this rdd partition and return an iterator.

source

RDD.AbstractRDDModule.dependencies — Method.

dependencies(rdd::AbstractRDD, partition::Int)::AbstractVector{Dependency}

Returns the list of parent rdds partitions partition of rdd depends on with the type of the dependency.

source

RDD.AbstractRDDModule.iterator — Method.

iterator(rdd::AbstractRDD, partition::Int, parentiters::AbstractVector)

Returns an iterator on a partition of the rdd given parent partition(s) iterators. This method must be implemented by subclasses of AbstractRDD.

source

RDD.AbstractRDDModule.partitioner — Method.

partitioner(rdd::AbstractRDD)::Union{AbstractPartitioner, Nothing}

Returns the partitioner of this rdd. Default implementation returns nothing.

source

RDD.AbstractRDDModule.partitions — Method.

partitions(rdd::AbstractRDD)::Int64

Returns the number of partitions of the given rdd. This method must be implemented by subclasses of AbstractRDD. The default implementation raises an error.

source

RDD.AbstractRDDModule.preferredlocations — Method.

preferredlocations(rdd::AbstractRDD, partition::Int)::AbstractVector{Int}

Returns the list of workers where it is preferable to compute this partition. The list is ordered by preference.

source