ParallelCollectionRDD

ParallelCollectionRDD

RDD.ParallelCollectionRDDModule.ParallelCollectionPartition

RDD.ParallelCollectionRDDModule.ParallelCollectionPartitionIterator

RDD.ParallelCollectionRDDModule.ParallelCollectionRDD

RDD.ParallelCollectionRDDModule.ParallelCollectionRDD

RDD.ParallelCollectionRDDModule.ParallelCollectionRDD

RDD.ParallelCollectionRDDModule.ParallelCollectionPartition — Type.

struct ParallelCollectionPartition{T} <: AbstractPartition{T}

Type of a partition of ParallelCollectionRDD

source

RDD.ParallelCollectionRDDModule.ParallelCollectionPartitionIterator — Type.

struct ParallelCollectionPartitionIterator{T} <: AbstractPartitionIterator{T}

Type of the iterator of a ParallelCollectionPartition

source

RDD.ParallelCollectionRDDModule.ParallelCollectionRDD — Type.

struct ParallelCollectionRDD{T} <: AbstractRDD{T}

A subclass of AbstractRDD representing distributed in-memory collection.

source

RDD.ParallelCollectionRDDModule.ParallelCollectionRDD — Method.

ParallelCollectionRDD{T}(col::AbstractVector{T}) where {T}

Create a ParallelCollectionRDD from a vector. The number of partitions will be the number of workers of a julia cluster as returned by workers(). If there is less than 2 workers, 2 partitions will be created.

source

RDD.ParallelCollectionRDDModule.ParallelCollectionRDD — Method.

ParallelCollectionRDD{T}(col::AbstractVector{T}, numpart::Int) where {T}

Create a ParallelCollectionRDD from a vector with numpart partitions.

source

RDD.AbstractRDDModule.iterator — Method.

iterator(
    rdd::ParallelCollectionRDD{T}, 
    numpart::Int,
    parent_iters::AbstractVector{AbstractPartitionIterator}
)::ParallelCollectionPartitionIterator{T} where {T}

Implementation of iterator for ParallelCollectionRDD.

source

RDD.AbstractRDDModule.partitions — Method.

partitions(
        rdd:: ParallelCollectionRDD{T}
    )::AbstractVector{ParallelCollectionPartition{T}} where {T}

Implementation of partitions for ParallelCollectionRDD.

source